pyproject.toml
[project]
name = "comfyui-geometrypack"
description = "CGAL - Blender - IGL - Paraview - VTK. Geometry processing nodes for ComfyUI - mesh manipulation, remeshing, hole filling, UV unwrapping, and 3D visualization"
version = "0.1.1"
license = { text = "GPL-3.0-or-later" }
classifiers = [
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "trimesh>=3.15.0",
    "libigl>=2.6.1",
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "matplotlib>=3.5.0",
    "Pillow>=9.0.0",
    "pymeshlab>=2022.2",
    "cgal>=0.9.0",
    "point-cloud-utils>=0.30.0",
    "mesh-to-sdf>=0.0.14",
    "xatlas>=0.0.11",
    "PyNanoInstantMeshes",
    "requests>=2.25.0",
    "tqdm>=4.60.0",
    "rtree",
    "fast-simplification>=0.1.5",
    "skeletor>=1.2.0"
]

[project.urls]
Repository = "https://github.com/PozzettiAndrea/ComfyUI-GeometryPack"
Documentation = "https://github.com/PozzettiAndrea/ComfyUI-GeometryPack/blob/main/README.md"

[tool.comfy]
PublisherId = "pznodes"
DisplayName = "ComfyUI-GeometryPack"
Icon = ""

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
